Implement new progressive darkening logic
The previous approach of reducing the RGB colours by 10% would
never reach 0 in 10 clicks (i.e. 0.9^10 ~= 0.35). Instead,
this new approach takes the largest of the RGB values  and divides
by 10. This value is then stored and used to reduce the color by
that amount, each time that particular square is clicked.